Create Commands.md
Creation of Commands.md in `docs/`. I've also added the four commands we currently have.
If anyone has anything to add on to this or if I made a typo somewhere, feel free to tell me so :)

I also recommend that in the future, once we get more commands, we separate them per category.