Fix Profile Command (See Desc)
Fixed an issue with the profile command.
The issue was caused due to the DB check for an existing user. It was comparing the Guild ID's which when you're using the command from a different server, would change.
This in turn would cause the code to jump into the `Store()` method rather than the `Update()` method.

Removing the Guild ID check fixed the issue. It's still safe too as User Id's are unique.